Split up quiver module
