+++ frontend/lib/src/components/elements/ArrowTable/ArrowTable.tsx
@@ -19,6 +19,7 @@ import React, { ReactElement } from "react"
 import range from "lodash/range"
 
 import { Quiver } from "@streamlit/lib/src/dataframes/Quiver"
+import { format as formatArrowCell } from "@streamlit/lib/src/dataframes/arrowFormatting"
 
 import {
   StyledEmptyTableCell,
@@ -103,7 +104,7 @@ function generateTableCell(
   } = table.getCell(rowIndex, columnIndex)
 
   const formattedContent =
-    displayContent || Quiver.format(content, contentType, field)
+    displayContent || formatArrowCell(content, contentType, field)
 
   const { headerColumns } = table.dimensions
   const cellDataType =
