+++ frontend/lib/src/components/widgets/DataFrame/columns/SelectboxColumn.ts
@@ -17,7 +17,8 @@
 import { GridCell, GridCellKind } from "@glideapps/glide-data-grid"
 import { DropdownCellType } from "@glideapps/glide-data-grid-cells"
 
-import { Quiver } from "@streamlit/lib/src/dataframes/Quiver"
+import { getTypeName } from "@streamlit/lib/src/dataframes/arrowTypeUtils"
+
 import {
   isNullOrUndefined,
   notNullOrUndefined,
@@ -53,8 +54,7 @@ function SelectboxColumn(props: BaseColumnProps): BaseColumn {
   const parameters = mergeColumnParameters(
     // Default parameters:
     {
-      options:
-        Quiver.getTypeName(props.arrowType) === "bool" ? [true, false] : [],
+      options: getTypeName(props.arrowType) === "bool" ? [true, false] : [],
     },
     // User parameters:
     props.columnTypeOptions
